Move after_tests to Tools
